feat(promo): Refactor CompanyPromoDetailView to RetrieveUpdateAPIView
Refactor CompanyPromoDetailView from a generic APIView to DRF’s
RetrieveUpdateAPIView in order to leverage built-in retrieve/update
behavior and simplify the implementation.

Key changes:
- Use `RetrieveUpdateAPIView` as the base class to get object lookup,
  serialization, validation and response handling for free
- Remove manual `get()` and `patch()` methods
- Add `business.permissions.IsPromoOwner` to `permission_classes` for
  declarative authorization
- Define `queryset = Promo.objects.with_related().only(...).for_company()`
  on the view to centralize and optimize data access
- Implement `with_related()` and `for_company(user)` on the Promo
  queryset/manager for efficient prefetching and filtering
- Drop unused imports and old helper code

This change reduces boilerplate, improves consistency with DRF
idioms, and makes permission checks and queryset logic easier to
maintain.
